chore: implement all suggested changes
